Fixed C# not working in builds with CompiledFromBlueprint flag set.
This commit will break current Blueprints, so unfortunately you have to remake the Blueprint classes or do manual redirects in DefaultEngine.ini. Just add _C at the end of every class that has a BP child, like this:

[CoreRedirects]
+ClassRedirects=(OldName="/Script/UnrealSharp.MyCSharpClass",NewName="/Script/UnrealSharp.MyCSharpClass_C")

The issue is that the C# classes and UBlueprints exists in the same package now, which causes name conflicts across the allocated UObjects in that package, but we need the dummy Blueprint UObject to have the class name without the suffix, so it looks correct in editor tools.

So we’ll do "_C" on the class UObject just like how Epic does it. Unfortunately, we can't let the class own the UBlueprint, like we have it now because then the linker doesn’t find it if we have CompiledFromBlueprint, which we need for some K2 events.