Add an extra check for the dotnet path on Nix.
Essentially, if i just give it the raw $DOTNET_PATH as explained by the NixOS manual, it complains that it can't find the dotnet executable. I'm unsure if this is due to symlinks or something, but there's an additional /bin directory in the Nix package for dotnet, and that seems to be the proper executable to run.