Initial commit of replacing weaver with incremental source gen
This commit includes a bit of tribal knowledge to get it working right now.

The biggest change for users is assemblies are now in Managed/DOT_NET_VERSION/ and Managed/DOT_NET_STANDARD_VERSION instead of just /Managed.
Also all types exposed to Unreal, needs to be partial, that includes all types + UProperties + UFunctions.

Currently it on-par with the weaver for Cropout and can be packaged to a shipping game. I have probably missed some things from the weaver.

The system needs some clean-up here and there.